refactor: Move Homepage's components into a new CreateDrawUI component
