+++ src/components/BulkAddLink/BulkAddLink.module.css
@@ -1,0 +1,13 @@
+.container {
+  text-align: center;
+  margin-top: 1rem;
+}
+
+.link {
+  color: #4CB4FF;
+  font-weight: bold;
+}
+.link:hover {
+  color: #3CA4EF;
+  text-decoration: underline;
+}
