+++ src/components/CreateDraw/SubmitValues/SubmitValues.connect.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import SubmitValues from './SubmitValues'
 import { connect } from 'react-redux'
-import * as valuesActions from '../../redux/features/draw/actions'
-import { RootState } from '../../redux/rootReducer'
+import * as valuesActions from '../../../redux/features/draw/actions'
+import { RootState } from '../../../redux/rootReducer'
 
 const mapStateToProps = (state: RootState) => ({
   values: state.draw.draw.values,
