+++ src/components/DrawResult/AnimatedValues/AnimatedValues.tsx
@@ -1,5 +1,5 @@
 import React, { useRef } from 'react'
-import Value from '../../Value/Value'
+import Value from '../../Shared/Value/Value'
 import Animation from '../../../redux/features/animation/models/Animation'
 import Values from '../../../redux/features/draw/models/Values'
 import styles from './AnimatedValues.module.css'
