+++ src/components/Shared/Value/Value.module.css
