+++ src/components/Shared/Value/Value.tsx
